requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy<2.0.0
pydantic>=2.0.0
